+++ src/components/shared/svgs.tsx
@@ -1,0 +1,36 @@
+import { type SVGProps, forwardRef } from 'react';
+
+import { cn } from '@/lib/utils';
+
+export const ThemeIcon = forwardRef<SVGSVGElement, SVGProps<SVGSVGElement>>(
+	({ className, ...props }, ref) => (
+		<svg
+			className={cn(
+				'size-4.5 transition-transform duration-300 ease-in-out',
+				className,
+			)}
+			fill="none"
+			ref={ref}
+			stroke="currentColor"
+			strokeLinecap="round"
+			strokeLinejoin="round"
+			strokeWidth="2"
+			viewBox="0 0 24 24"
+			xmlns="http://www.w3.org/2000/svg"
+			{...props}
+		>
+			<path
+				d="M0 0h24v24H0z"
+				fill="none"
+				stroke="none"
+			/>
+			<path d="M12 12m-9 0a9 9 0 1 0 18 0a9 9 0 1 0 -18 0" />
+			<path d="M12 3l0 18" />
+			<path d="M12 9l4.65 -4.65" />
+			<path d="M12 14.3l7.37 -7.37" />
+			<path d="M12 19.6l8.85 -8.85" />
+		</svg>
+	),
+);
+
+ThemeIcon.displayName = 'ThemeIcon';